Rename forwarder to get around pathing
Mysterious issue when running within docker, that because forwarder package and forwarder module had same name, couldn't import from package init file.   Got around this by renaming package to forwarderservice, which is more accurate anyway.
