--- forwarder/forwarder/queues/__init__.py
@@ -1,3 +1,0 @@
-from forwarder.queues.redis.redis_q import RedisQueue
-
-__all__ = ['RedisQueue']
