+++ forwarder/forwarderservice/queues/__init__.py
@@ -1,0 +1,3 @@
+from forwarderservice.queues.redis.redis_q import RedisQueue
+
+__all__ = ['RedisQueue']
